+++ crates/zng/src/drag_drop.rs
@@ -35,8 +35,8 @@
 //! See [`zng_ext_input::drag_drop`] and [`zng_wgt_input::drag_drop`] for the full drag&drop API.
 
 pub use zng_ext_input::drag_drop::{
-    DragDropData, DragHandle, SystemDragDropData, WeakDragHandle, DRAG_DROP, DRAG_END_EVENT, DRAG_HOVERED_EVENT, DRAG_START_EVENT,
-    DROP_EVENT,
+    DragDropData, DragEndArgs, DragHandle, DragHoveredArgs, DragStartArgs, DropArgs, SystemDragDropData, WeakDragHandle, DRAG_DROP,
+    DRAG_END_EVENT, DRAG_HOVERED_EVENT, DRAG_START_EVENT, DROP_EVENT,
 };
 
 pub use zng_wgt_input::drag_drop::{